ART-902 support for special characters
