+++ core/src/main/java/nl/nn/adapterframework/align/XmlAligner.java
@@ -349,5 +349,5 @@ public class XmlAligner extends XMLFilterImpl {
 		}
 		return null;
 	}
-	
+
 }
